requirements.txt
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.0
decorator==5.1.1
Django==4.1.1
django-environ==0.9.0
djangorestframework==3.13.1
environ==1.0
jsonschema==4.16.0
protobuf==4.21.7
psycopg2==2.9.3
pyrsistent==0.18.1
pytz==2022.2.1
sqlparse==0.4.2
validators==0.20.0
asgiref==3.5.2
attrs==22.1.0
bcrypt==4.0.0
boto3==1.24.96
botocore==1.27.96
decorator==5.1.1
Django==4.1.1
django-environ==0.9.0
djangorestframework==3.13.1
environ==1.0
jmespath==1.0.1
jsonschema==4.16.0
protobuf==4.21.7
psycopg2==2.9.3
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
validators==0.20.0
statsd==4.0.0
django-statsd-mozilla==0.4.0